cat: store buffer sizes as 'long long'
Use 'long long' instead of 'int' for buffer sizes to prevent overflow
from occurring for large buffer sizes.

These changes have been tested on the AMD Zen4 architecture.

+++ src/counter_analysis_toolkit/dcache.h
@@ -6,9 +6,9 @@
 #include "hw_desc.h"
 #include "params.h"
 
-int varyBufferSizes(int *values, double **rslts, double **counter, hw_desc_t *hw_desc, int line_size_in_bytes, float pages_per_block, int pattern, int latency_only, int mode, int ONT);
+int varyBufferSizes(long long *values, double **rslts, double **counter, hw_desc_t *hw_desc, long long line_size_in_bytes, float pages_per_block, int pattern, int latency_only, int mode, int ONT);
 int get_thread_count();
 void d_cache_driver(char* papi_event_name, cat_params_t params, hw_desc_t *hw_desc, int latency_only, int mode);
-int d_cache_test(int pattern, int max_iter, hw_desc_t *hw_desc, int stride_in_bytes, float pages_per_block, char* papi_event_name, int latency_only, int mode, FILE* ofp);
+int d_cache_test(int pattern, int max_iter, hw_desc_t *hw_desc, long long stride_in_bytes, float pages_per_block, char* papi_event_name, int latency_only, int mode, FILE* ofp);
 
 #endif
